Removed regexp and accounted using configuration. Added differentiation between good and bad dry-runs, formatted and output the dry-run to the caller, added proper snakemake version to project along with bioconda channel, vaulted use of PipelineOut class in create pipeline workflow

+++ routes/create/create.py
@@ -9,7 +9,7 @@ from motor.motor_asyncio import AsyncIOMotorCollection
 from db import get_database
 
 database = get_database()
-snakemake_pipelines = database["snakemake_pipeline"]
+snakemake_pipelines_collection = database["snakemake_pipeline"]
 
 # Check if git_url exists in database
 async def git_url_exists(git_url: str, collection: AsyncIOMotorCollection) -> bool:
@@ -18,28 +18,23 @@ async def git_url_exists(git_url: str, collection: AsyncIOMotorCollection) -> bo
         return True
     return False
 
-# CREATE
 # Add pipeline entry to database
 async def add_pipeline(
     pipeline: CreatePipeline,
     collection: AsyncIOMotorCollection,
-) -> PipelineOut:
+) -> None:
     
     try:
-        new_pipeline = pipeline.model_dump()
-        result = await collection.insert_one(new_pipeline)
-        new_pipeline['id'] = result.inserted_id
+        await collection.insert_one(pipeline.model_dump())
     except ValueError as error:
         await pipeline.delete_local()
         raise HTTPException(status_code=401, detail=str(error))
    
-    return PipelineOut(**new_pipeline)
-
 # Recieve a pipeline name, github url, list of output files, path to the snakefile, path to pipeline configurations, path to conda environment
-async def create_pipeline(data: CreatePipeline) -> PipelineOut:
+async def create_pipeline(data: CreatePipeline) -> None:
     try:
         pipeline = CreatePipeline(**data)
-        if await git_url_exists(pipeline.git_url, snakemake_pipelines):
+        if await git_url_exists(pipeline.git_url, snakemake_pipelines_collection):
             raise HTTPException(status_code=400, detail="Git url already exists in database")
 
     except KeyError as error:
@@ -56,17 +51,19 @@ async def create_pipeline(data: CreatePipeline) -> PipelineOut:
     await pipeline.validate_local_file_paths()
     
     # perform a dry run of the pipeline
-    try:
-        await pipeline.dry_run()
-    except Exception as e:
+    dry_run_status = await pipeline.dry_run()
+
+    # if dry-run contains unsucessful output throw exception
+    if "The order of jobs does not reflect the order of execution" not in dry_run_status:
         await pipeline.delete_local()
-        raise HTTPException(status_code=400, detail=(f"Error performing dry run: {e}"))
+        raise HTTPException(status_code=400, detail=(f"Error performing dry run: {dry_run_status}"))
 
     # add to database
-    pipeline_out = await add_pipeline(pipeline, snakemake_pipelines)
+    await add_pipeline(pipeline, snakemake_pipelines_collection)
 
     raise HTTPException(status_code=200, detail={"clone_status": "Pipeline cloned successfully",
                                                  "configuration_checks": "Pipeline passed configuration checks",
+                                                 "dry_run_status": str(dry_run_status),
                                                  "pipeline_directory": str(pipeline.fs_path),
-                                                 "pipeline_database_entry": pipeline_out.model_dump()})
+                                                 "pipeline_database_entry": pipeline.model_dump()})
